+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Linker.scala
@@ -17,7 +17,7 @@ import io.buoyant.namer._
 import io.buoyant.linkerd.telemeter.UsageDataTelemeterConfig
 import io.buoyant.telemetry._
 import io.buoyant.telemetry.admin.{AdminMetricsExportTelemeter, histogramSnapshotInterval}
-import java.net.InetSocketAddress
+import java.net.{InetAddress, InetSocketAddress}
 import scala.util.control.NoStackTrace
 
 /**
@@ -35,7 +35,7 @@ trait Linker {
 object Linker {
   private[this] val log = Logger()
 
-  private[this] val DefaultAdminAddress = new InetSocketAddress(9990)
+  private[this] val DefaultAdminAddress = new InetSocketAddress(InetAddress.getLoopbackAddress, 9990)
   private[this] val DefaultAdminConfig = AdminConfig()
 
   private[linkerd] case class Initializers(